--- storage/core/exporter.go
@@ -1,1 +1,0 @@
-package core
