+++ storage/core/rpcService/importer_test.go
@@ -1,4 +1,4 @@
-package core
+package rpcService
 
 import (
 	"io"
